requirements.txt
streamlit
ultralytics
torch
Pillow
